sync w/ main branch

+++ view/admin/kelola_ruang.php
@@ -13,10 +13,8 @@
             <input class="search-box" type="text" placeholder=" Pilih ID Ruang">
             <button class="search-button">Cari</button>
 
-            <button type="button" class="tambah" data-bs-toggle="modal" data-bs-target="#exampleModal"
-                data-bs-whatever="@mdo">
-                <i class="bi bi-person-add"></i>Tambah
-            </button>
+            <button type="button" class="tambah" data-bs-toggle="modal" data-bs-target="#exampleModal" data-bs-whatever="@mdo">
+                <i class="bi bi-person-add"></i>Tambah</button>
         </div>
         <table class="table " style="table-layout: auto;">
             <thead>
@@ -35,31 +33,28 @@
                 if (!empty($query)) {
                     foreach ($query as $row) {
                 ?>
-                <tr>
-                    <td scope="row"><?= $no++; ?></td>
-                    <td><?= $row['NamaRuang']; ?></td>
-                    <!-- <td><?= $row['JenisRuang']; ?></td> -->
-                    <td><?= $row['DeskripsiRuang']; ?></td>
-                    <td><?= $row['Lantai']; ?></td>
-                    <td>
-                        <!-- <a href="index.php?page=anggota/edit&id=<?php echo $row['RuangID']; ?>"
+                        <tr>
+                            <td scope="row"><?= $no++; ?></td>
+                            <td><?= $row['NamaRuang']; ?></td>
+                            <!-- <td><?= $row['JenisRuang']; ?></td> -->
+                            <td><?= $row['DeskripsiRuang']; ?></td>
+                            <td><?= $row['Lantai']; ?></td>
+                            <td>
+                                <!-- <a href="index.php?page=anggota/edit&id=<?php echo $row['RuangID']; ?>"
                             class="btn btn-warning btn-xs"><i class="fa fa-pencil-square-o" aria-hidden="true"></i>
                             Edit</a> -->
-                        <a role="button" class="btn btn-warning" data-bs-toggle="modal" data-bs-target="#editModal"
-                            data-bs-whatever="@mdo">Edit</a>
-                        <a href="fungsi/hapus.php?anggota=hapus&id=<?php echo $row['RuangID']; ?>"
-                            onclick="javascript:return confirm('Hapus Data Anggota ?');"
-                            class="btn btn-danger btn-xs"><i class="fa fa-trash-o" aria-hidden="true"></i> Hapus</a>
-                    </td>
-                    <?php
+                                <a role="button" class="btn btn-warning" data-bs-toggle="modal" data-bs-target="#editModal" data-bs-whatever="@mdo">Edit</a>
+                                <a href="fungsi/hapus.php?anggota=hapus&id=<?php echo $row['RuangID']; ?>" onclick="javascript:return confirm('Hapus Data Anggota ?');" class="btn btn-danger btn-xs"><i class="fa fa-trash-o" aria-hidden="true"></i> Hapus</a>
+                            </td>
+                        <?php
                     }
                 } else {
                         ?>
-                    <td colspan="4">Tidak Ada Data Tersedia</td>
+                        <td colspan="4">Tidak Ada Data Tersedia</td>
                     <?php
                 }
                     ?>
-                </tr>
+                        </tr>
             </tbody>
         </table>
     </div>
@@ -69,8 +64,7 @@
 </main>
 
 <!-- TAMBAH RUANG -->
-<div class="modal fade" id="exampleModal" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog"
-    aria-labelledby="modalTitleId" aria-hidden="true">
+<div class="modal fade" id="exampleModal" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog" aria-labelledby="modalTitleId" aria-hidden="true">
     <div class="modal-dialog">
         <div class="modal-content">
             <div class="modal-header">
@@ -115,10 +109,8 @@
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <button type="button" class="btn btn-warning" data-bs-dismiss="modal" aria-hidden="true"><i
-                            class="bi bi-x-lg"></i>Simpan</button>
-                    <button type="submit" class="btn btn-danger" aria-hidden="true"><i
-                            class="bi bi-floppy"></i>Hapus</button>
+                    <button type="button" class="btn btn-warning" data-bs-dismiss="modal" aria-hidden="true"><i class="bi bi-x-lg"></i>Simpan</button>
+                    <button type="submit" class="btn btn-secondary" aria-hidden="true"><i class="bi bi-floppy"></i>Batal</button>
                 </div>
             </form>
         </div>
@@ -127,8 +119,7 @@
 <!-- END TAMBAH RUANG -->
 
 <!-- Edit Ruang -->
-<div class="modal fade" id="editModal" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog"
-    aria-labelledby="modalTitleId" aria-hidden="true">
+<div class="modal fade" id="editModal" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog" aria-labelledby="modalTitleId" aria-hidden="true">
     <div class="modal-dialog">
         <div class="modal-content">
             <div class="modal-header">
@@ -173,10 +164,9 @@
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <button type="button" class="btn btn-warning" data-bs-dismiss="modal" aria-hidden="true"><i
-                            class="bi bi-x-lg"></i>Ubah</button>
-                    <button type="submit" class="btn btn-danger" aria-hidden="true"><i
-                            class="bi bi-floppy"></i>Hapus</button>
+                    <button type="button" class="btn btn-warning" data-bs-dismiss="modal" aria-hidden="true"><i class="bi bi-x-lg"></i>Ubah</button>
+                    <!-- <button type="submit" class="btn btn-danger" aria-hidden="true"><i
+                            class="bi bi-floppy"></i>Hapus</button> -->
                 </div>
             </form>
         </div>
@@ -184,8 +174,7 @@
 </div>
 <!-- End Edit Ruang -->
 
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous">
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous">
 </script>
 </body>
